Refactoring interact (shared by both evo1 and evo2)
