+++ engine/seq/__init__.py
@@ -1,8 +1,9 @@
 from engine.seq.base import SeqBase, SeqList, SeqOptional
+from engine.seq.interact import SeqAttack, SeqTapDirection
 from engine.seq.log import SeqDebug, SeqLog
 from engine.seq.sequencer import SequencerEngine
 from engine.seq.start import EvolandStartGame
-from engine.seq.time import SeqDelay, wait_seconds
+from engine.seq.time import SeqDelay, SeqMashDelay, wait_seconds
 
 __all__ = [
     "EvolandStartGame",
@@ -10,8 +11,11 @@ __all__ = [
     "SeqDebug",
     "SeqLog",
     "SeqDelay",
+    "SeqMashDelay",
     "wait_seconds",
     "SeqBase",
     "SeqList",
     "SeqOptional",
+    "SeqTapDirection",
+    "SeqAttack",
 ]
