+++ evo1/seq/crystal_cavern.py
@@ -2,9 +2,8 @@ import logging
 
 from control import evo_ctrl
 from engine.mathlib import Facing, Vec2
-from engine.seq import SeqList
+from engine.seq import SeqList, SeqMashDelay
 from evo1.atb import EncounterID, SeqATBCombat, SeqATBmove2D
-from evo1.interact import SeqMashDelay
 from evo1.maps import GetAStar
 from evo1.memory import MapID, get_zelda_memory
 from evo1.move2d import (
