+++ evo1/seq/edel_vale.py
@@ -1,10 +1,9 @@
 from engine.mathlib import Box2, Facing, Vec2
-from engine.seq import SeqList
+from engine.seq import SeqAttack, SeqList
 from evo1.combat import SeqKnight2D, SeqMove2DClunkyCombat
 from evo1.maps import GetAStar
 from evo1.memory import MapID
 from evo1.move2d import (
-    SeqAttack,
     SeqGrabChest,
     SeqGrabChestKeyItem,
     SeqMove2D,
